Move Master Site class to its own file to abide to the PHPCS standard.
